Rename template controllers and views :
news/entry -> news-details
news/index -> news-index
event/entry -> event-details
event/index -> event-list
