+++ templates/boilerplate/template/event-details.mustache
