+++ templates/boilerplate/template/news-details.mustache
