+++ templates/boilerplate/template/news-list.mustache
